+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- introduce I18N_PO_DATA_BY_LANGUAGE to get raw gettext data
+
 ### Changes
 
 ### Deprecated
@@ -17,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- BREAKING: remove `concept-schemes.json` from VOCABULARY_JSON_BY_NAME
 - BREAKING: fixed typo in the technical accessibility concept scheme identifier
 - BREAKING: fixed copy-pasta in merged-resource-series identifier
 - fixed typos and grammar in descriptions: comma before "that", punctuation, whitespaces